Improved compaction

+++ src/main/java/com/arcadedb/engine/PBucket.java
@@ -50,7 +50,7 @@ public class PBucket extends PPaginatedFile {
   }
 
   @Override
-  protected long getPageSize() {
+  protected int getPageSize() {
     return PAGE_SIZE;
   }
